fix: casting data type of Balance prop inside AtlasCommonResponse to long

+++ Platipus.Wallet.Api/Application/Requests/Wallets/Atlas/AtlasBetRequest.cs
@@ -40,7 +40,8 @@ public sealed record AtlasBetRequest(
                 return walletResult.ToAtlasFailureResult<AtlasCommonResponse>();
             
             var response = new AtlasCommonResponse(
-                walletResult.Data?.Currency!, walletResult.Data!.Balance * 100, 
+                walletResult.Data?.Currency!, 
+                (long)walletResult.Data!.Balance * 100, 
                 walletResult.Data.UserId.ToString());
 
             return AtlasResultFactory.Success(response);
